big update lots of changes to dungeon crawler and trader game

+++ Code/GenrpgClient/Assets/Scripts/Buildings/WeightedBuildingMaterial.cs
@@ -7,7 +7,7 @@ namespace Assets.Scripts.Buildings
     [Serializable]
     public class WeightedBuildingMaterial
     {
-        public float Weight;
+        public double Weight;
         public Material Mat;
         public List<Color> ColorTargets = new List<Color>();
     }
